Move CommandFactory to ObjectModel
