+++ samples/XCT.Sample/Pages/Effects/TouchEffectPage.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using System.Windows.Input;
-using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.ObjectModel;
 using Xamarin.Forms.Xaml;
 
 namespace Xamarin.CommunityToolkit.Sample.Pages.Effects
