+++ samples/XCT.Sample/Pages/TestCases/TouchEffectButtonPage.xaml.cs
@@ -1,5 +1,5 @@
 ﻿using System.Windows.Input;
-using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.ObjectModel;
 
 namespace Xamarin.CommunityToolkit.Sample.Pages.TestCases
 {
