+++ samples/XCT.Sample/ViewModels/AboutViewModel.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using Octokit;
-using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.ObjectModel;
 using Xamarin.Essentials;
 
 namespace Xamarin.CommunityToolkit.Sample.ViewModels
