+++ samples/XCT.Sample/ViewModels/Base/BaseGalleryViewModel.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Input;
-using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.ObjectModel;
 using Xamarin.CommunityToolkit.Sample.Models;
 
 namespace Xamarin.CommunityToolkit.Sample.ViewModels
