+++ samples/XCT.Sample/ViewModels/Behaviors/EventToCommandBehaviorViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using System.Windows.Input;
-using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.ObjectModel;
 
 namespace Xamarin.CommunityToolkit.Sample.ViewModels.Behaviors
 {
