+++ samples/XCT.Sample/ViewModels/Views/StateLayoutViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using System.Windows.Input;
-using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.ObjectModel;
 using Xamarin.CommunityToolkit.UI.Views;
 
 namespace Xamarin.CommunityToolkit.Sample.ViewModels.Views
